pyproject.toml
[project]
name = "telegram-bot-for-events"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.25.0",
    "aiosqlite>=0.22.1",
    "asyncpg>=0.31.0",
    "bcrypt==4.0.1",
    "fastapi>=0.129.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "passlib==1.7.4",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]==3.3.0",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
